requirements.txt
absl-py==0.11.0
amqp==2.5.1
aniso8601==7.0.0
appdirs==1.4.3
arabic-reshaper==2.1.1
argon2-cffi==20.1.0
asgiref==3.3.1
astunparse==1.6.3
async-generator==1.10
attrs==19.1.0
Babel==2.7.0
backcall==0.2.0
backports.csv==1.0.7
beautifulsoup4==4.9.3
bert-extractive-summarizer==0.5.1
billiard==3.6.1.0
black==19.3b0
bleach==3.2.1
blis==0.2.4
boto3==1.16.39
botocore==1.19.39
bs4==0.0.1
CacheControl==0.12.5
cachetools==3.1.1
catalogue==1.0.0
celery==4.3.0
certifi==2020.12.5
cffi==1.13.2
chardet==3.0.4
Click==7.0
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
convertdate==2.3.0
cryptography==2.8
cycler==0.10.0
cymem==2.0.4
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.1
Django==3.1.4
django-celery-beat==1.5.0
django-cors-headers==3.6.0
django-debug-toolbar==2.0
django-filter==2.2.0
django-graphql-geojson==0.1.4
django-graphql-jwt==0.3.1
django-import-export==1.2.0
django-simple-history==2.10.0
django-tenants==3.2.1
django-timezone-field==3.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djongo==1.3.3
dnspython==2.0.0
docutils==0.15.2
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-2.3.1/en_core_web_md-2.3.1.tar.gz
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints==0.3
et-xmlfile==1.0.1
fastapi==0.61.2
filelock==3.0.12
firebase-admin==4.5.1
future==0.18.2
gast==0.3.3
gensim==3.8.3
geographiclib==1.50
geopy==2.1.0
google-api-core==1.25.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-cloud-core==1.5.0
google-cloud-firestore==2.0.2
google-cloud-speech==2.0.1
google-cloud-storage==1.35.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
grpcio==1.33.2
gunicorn==20.0.4
h11==0.11.0
h5py==2.10.0
holidays==0.10.3
html5lib==1.1
httplib2==0.18.1
idna==2.10
importlib-metadata==3.7.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kombu==4.6.4
korean-lunar-calendar==0.2.1
libcst==0.3.16
mango-jwt==1.3.2
Markdown==3.3.3
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
more-itertools==7.2.0
mpmath==1.1.0
msgpack==0.6.1
murmurhash==1.0.4
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
neuralcoref==4.0
nltk==3.2.5
notebook==6.1.5
numpy==1.18.5
oauthlib==3.1.0
odfpy==1.4.0
opencv-python==4.4.0.46
openpyxl==2.6.3
opt-einsum==3.3.0
ortools==8.2.8710
overrides==3.1.0
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
passlib==1.7.4
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plac==0.9.6
preshed==2.0.1
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.8
proto-plus==1.13.0
protobuf==3.14.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pydantic==1.7.2
pydot==1.4.1
Pygments==2.7.2
PyJWT==1.7.1
PyMeeus==0.3.7
pymongo==3.11.2
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
python-bidi==0.4.2
python-crontab==2.3.8
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2020.4
PyYAML==5.1.2
pyzmq==20.0.0
qtconsole==4.7.7
QtPy==1.9.0
regex==2020.11.13
reportlab==3.5.62
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
Rx==1.6.1
s3transfer==0.3.3
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.4
Send2Trash==1.5.0
sentencepiece==0.1.91
singledispatch==3.4.0.3
six==1.15.0
sklearn==0.0
sklearn-contrib-py-earth==0.1.0+0.gb209d19.dirty
smart-open==3.0.0
soupsieve==2.0.1
spacy==2.1.3
sqlparse==0.2.4
srsly==1.0.4
starlette==0.13.6
tablib==0.13.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
tensorflow-hub==0.10.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
thinc==7.0.8
threadpoolctl==2.1.0
tokenizers==0.9.3
toml==0.10.0
torch==1.7.0
tornado==6.1
tqdm==4.52.0
traitlets==5.0.5
transformers==2.2.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.26.2
uvicorn==0.12.3
vine==1.3.0
virtualenv==20.2.2
vt-agronomy==0.0.4
vt-core==0.0.5
wasabi==0.8.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wget==3.2
whitenoise==5.2.0
widgetsnbextension==3.5.1
wrapt==1.12.1
xhtml2pdf==0.2.5
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0